Rating Widget Improvements.
Dramatically improve the performance of the rating widget by moving the rendering code to a separate widget-agnostic class that keeps a cache of previously rendered states. This was necessary to make the code more robust, and also to improve the performace of the cell renderer, since the previous implementation slowed down scrolling through the song list.