ViewWrapper: add the ability to cancel searches.
This has serious performance benefits when a user is searching in a huge library. Previously, the search displayed the results only after the previous searches (produced by the previous letters they typed) had been processed (a waste of processor time indeed). Now, after a new key is pressed, all the previous searches are cancelled, resulting in a performance boost and smoother visual response.