Change the queue mechanism to make it less confusing
Fix #2

Now, when you double-click on a song, it is played and other songs of
the current list are added to the queue.

Also improved the style of some parts of the code, and changed many
Gee.HashMap<int, Media> to ArrayList<int, Media> (they were acting as
arrays, I don't know why they were maps).