LibraryWindow: Use action accels for shortcuts (#338)
Use action accels instead of manually grabbing keycodes from Gdk

This is a little more verbose, but using actions here is probably the right thing to do especially with regard to view switching. In the future we can set these actions as disabled when they're unavailable instead of relying on a widget's sensitivity.